requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
gunicorn==20.1.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
Werkzeug==2.3.4
zipp==3.15.0
